parameters reader refactor (CPUTEST)
