+++ src/framework/tests/parameters.cpp
@@ -1,4 +1,4 @@
-#include "framework/parameters.h"
+#include "framework/parameters/parameters.h"
 
 #include "defaults.h"
 #include "enums.h"
